pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "reactor-digital-twin"
version = "0.1.0"
description = "Digital Twin Batch Reactor simulation with OPC UA interface"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pyomo>=6.7",
    "asyncua>=1.1",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-timeout>=2.0",
]
[project.scripts]
reactor = "reactor.__main__:main_sync"
opc-tool = "opc_tool.__main__:main_sync"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-m 'not performance'"
markers = [
    "performance: optional performance/benchmark tests (run with: pytest -m performance)",
]
